Remove start / stop support as it is much more tricky as I thought.
The current implementation leaves the graphics context in "flush pending" state.
An obvious solution would be to block on the main thread until there is no
render pending, but this seems to be prone to deadlocks (with Graphics2D::Flush
waiting for the main thread to unblock). Other  possible solutions would be
replacing the graphics context when then renderer is restarted or notifying the
thread via a message and polling it until it has terminated, yielding control
between poll cycles. However, I've decided this exceeds the scope of this demo
:)